# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/artworks": {
            "get": {
                "description": "分页获取作品列表，支持过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "获取作品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "艺术家",
                        "name": "artist",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ArtworkResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的艺术作品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "创建作品",
                "parameters": [
                    {
                        "description": "作品信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArtworkCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ArtworkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/artworks/category/{category}": {
            "get": {
                "description": "获取指定分类的作品列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "按分类获取作品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/artworks/random": {
            "get": {
                "description": "随机获取指定数量的作品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "随机获取作品",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "艺术家",
                        "name": "artist",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ArtworkResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/artworks/upload": {
            "post": {
                "description": "上传图片到 CDN 并同时创建艺术作品记录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "上传文件并创建作品",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "作品标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "艺术家名称",
                        "name": "artist",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "作品描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "头像URL",
                        "name": "avatar_url",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/artworks/{id}": {
            "get": {
                "description": "获取指定ID的作品详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "获取作品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ArtworkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定ID的作品信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "更新作品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "作品信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArtworkUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ArtworkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定ID的作品",
                "tags": [
                    "Artwork"
                ],
                "summary": "删除作品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/artworks/{id}/bookmark": {
            "post": {
                "description": "增加作品的收藏数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "增加收藏数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/artworks/{id}/like": {
            "post": {
                "description": "增加作品的点赞数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "增加作品点赞数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/artworks/{id}/unbookmark": {
            "post": {
                "description": "减少作品的收藏数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "取消收藏",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/artworks/{id}/unlike": {
            "post": {
                "description": "减少作品的点赞数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artwork"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ArtworkCreateRequest": {
            "type": "object",
            "required": [
                "artist",
                "title",
                "url"
            ],
            "properties": {
                "artist": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "cdn_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "request_id": {
                    "description": "请求追踪ID",
                    "type": "string"
                },
                "timestamp": {
                    "description": "时间戳(毫秒)",
                    "type": "integer"
                }
            }
        }
    }
}